requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmdstanpy==1.1.0
comm==0.1.3
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
dill==0.3.5.1
einops==0.6.1
ephem==4.1.4
executing==1.2.0
fastavro==0.23.3
filelock==3.12.0
fire==0.5.0
flatbuffers==1.12
fonttools==4.39.3
gast==0.4.0
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
hijri-converter==2.3.1
holidays==0.24
idna==3.4
importlib-metadata==6.4.1
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.12.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
libclang==16.0.0
LunarCalendar==0.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
numpy==1.24.2
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.4.4
parso==0.8.3
patool==1.12
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.2.0
pmdarima==2.0.3
prompt-toolkit==3.0.38
prophet==1.1
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.0
PyMeeus==0.5.12
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.8.1
setuptools==66.0.0
setuptools-git==1.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
xlrd==2.0.1
zipp==3.15.0
